Source drop October 2025

+++ machines/hc12/machine.c
@@ -789,7 +789,7 @@ static void gen_pop(FILE *f,long l)
     BSET(regs_modified,iy);
 #endif
   }else{
-    emit(f,"\tleas\t%u,%s\n",SGN16(l),regnames[sp]);
+    emit(f,"\tleas\t%ld,%s\n",SGN16(l),regnames[sp]);
   }
   pop(l);
 }
@@ -2855,9 +2855,9 @@ void gen_code(FILE *f,struct IC *fp,struct Var *v,zmax offset)
 	    pd=(CPU!=6812&&cnt<=255)?1:2;
 	  }
 	  if(CPU!=6812&&cnt<=255)
-	    emit(f,"\tldb\t#%lu\n",cnt);
+	    emit(f,"\tldb\t#%u\n",(unsigned int)cnt);
 	  else
-	    emit(f,"\tldd\t#%lu\n",cnt);
+	    emit(f,"\tldd\t#%u\n",(unsigned int)cnt);
 	  cc=0;
 #if 0
 	  if(CPU!=6812&&((!regsa[iu]&&regs[iu])||drel)){
